Cover the link-creation happy path, count fetches in the guard test
Nothing exercised select -> create -> 201 -> toast -> re-render, so a
transposed argument in createLink()'s guard clause would have gone
unnoticed. Add that test, which doubles as proof the selection resets
after a create.

The self-link test also had no teeth: assertSelectorWillNotBeVisible on
#linksToastSuccess is satisfied on the first poll, since the toast is
already hidden at t=0, and passed with the guard deleted. Intercept
window.fetch and assert zero calls instead.

Both live in the createLink flow, so fold the guard file into
OffcanvasLinksCreateTest rather than adding a 20th subclass of
AbstractBrowserTestCase, which trips PHPMD's NumberOfChildren.
